+++ mountpoint-s3/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased (v1.21.0)
 
 * Change default logging level from WARN to INFO to improve visibility of important operational messages. ([#1605](https://github.com/awslabs/mountpoint-s3/pull/1605))
+* Fix readdir race condition that caused crashes during concurrent file deletions. ([#1648](https://github.com/awslabs/mountpoint-s3/pull/1648))
 
 ## v1.20.0 (Sep 12, 2025)